moved classes/interfaces from package j9adapters into class TextFlowExt as private nested classes/interfaces
