+++ richtextfx/build.gradle
@@ -73,8 +73,6 @@ javadoc {
         // resolve links to Flowless
         'https://fxmisc.github.io/flowless/javadoc/0.5/'
     ]
-
-    exclude('**/j9adapters/**')
 }
 
 // use test logging.
