+++ richtextfx/src/main/java/org/fxmisc/richtext/TextFlowExt.java
@@ -1,20 +1,19 @@
 package org.fxmisc.richtext;
 
 import static org.fxmisc.richtext.model.TwoDimensional.Bias.*;
-
+import java.lang.reflect.Array;
+import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.lang.reflect.Proxy;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 
-import javafx.scene.control.IndexRange;
-import org.fxmisc.richtext.j9adapters.GenericIceBreaker;
-import org.fxmisc.richtext.j9adapters.HitInfo;
-import org.fxmisc.richtext.j9adapters.RectBounds;
-import org.fxmisc.richtext.j9adapters.TextLayout;
-import org.fxmisc.richtext.j9adapters.TextLine;
 import org.fxmisc.richtext.model.TwoLevelNavigator;
 
+import javafx.scene.control.IndexRange;
 import javafx.scene.shape.PathElement;
 import javafx.scene.text.TextFlow;
 import javafx.scene.shape.LineTo;
@@ -202,5 +201,149 @@ class TextFlowExt extends TextFlow {
 
     private TextLayout textLayout() {
         return GenericIceBreaker.proxy(TextLayout.class, invoke(mGetTextLayout, this));
+    }
+
+    /* ********************************************************************** *
+     *                                                                        *
+     * GenericIceBreaker                                                      *
+     *                                                                        *
+     * ********************************************************************** */
+
+    private static class GenericIceBreaker implements InvocationHandler {
+        private final Object delegate;
+
+        private GenericIceBreaker(Object delegate) {
+            this.delegate = delegate;
+        }
+
+        @Override
+        public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+            Method delegateMethod = getDeclaredMethod(delegate.getClass(), method.getName(), method.getParameterTypes());
+
+            Object delegateMethodReturn = null;
+            try {
+                delegateMethodReturn = delegateMethod.invoke(delegate, args);
+            } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
+                throw new RuntimeException("problems invoking " + method.getName());
+            }
+            if (delegateMethodReturn == null) {
+                return null;
+            }
+
+            if (method.getReturnType().isArray()) {
+                if (method.getReturnType().getComponentType().isInterface()
+                        && !method.getReturnType().getComponentType().equals(delegateMethod.getReturnType().getComponentType())) {
+
+                    int arrayLength = Array.getLength(delegateMethodReturn);
+                    Object retArray = Array.newInstance(method.getReturnType().getComponentType(), arrayLength);
+                    for (int i = 0; i < arrayLength; i++) {
+                        Array.set(retArray,
+                                i,
+                                proxy(
+                                        method.getReturnType().getComponentType(),
+                                        Array.get(delegateMethodReturn, i)));
+                    }
+
+                    return retArray;
+                }
+            }
+
+            if (method.getReturnType().isInterface()
+                    && !method.getReturnType().equals(delegateMethod.getReturnType())) {
+                return proxy(method.getReturnType(), delegateMethodReturn);
+            }
+
+            return delegateMethodReturn;
+        }
+
+        @SuppressWarnings("unchecked")
+        static <T> T proxy(Class<T> iface, Object delegate) {
+            return (T) Proxy.newProxyInstance(
+                    iface.getClassLoader(),
+                    new Class[]{iface},
+                    new GenericIceBreaker(delegate));
+        }
+
+        private static final HashMap<MethodCacheKey, Method> declaredMethodCache = new HashMap<>();
+
+        private static synchronized Method getDeclaredMethod(Class<?> cls, String name, Class<?>... paramTypes)
+                throws NoSuchMethodException, SecurityException
+        {
+            MethodCacheKey methodCacheKey = new MethodCacheKey(cls, name, paramTypes);
+
+            Method m = declaredMethodCache.get(methodCacheKey);
+            if (m == null) {
+                m = cls.getDeclaredMethod(name, paramTypes);
+                m.setAccessible(true);
+                declaredMethodCache.put(methodCacheKey, m);
+            }
+            return m;
+        }
+    }
+
+    private static class MethodCacheKey {
+        final Class<?> cls;
+        final String name;
+        final Class<?>[] paramTypes;
+
+        MethodCacheKey(Class<?> cls, String name, Class<?>... paramTypes) {
+            this.cls = cls;
+            this.name = name;
+            this.paramTypes = paramTypes;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            if (!(obj instanceof MethodCacheKey))
+                return false;
+
+            MethodCacheKey key2 = (MethodCacheKey) obj;
+            return cls == key2.cls && name.equals(key2.name) && Arrays.equals(paramTypes, key2.paramTypes);
+        }
+
+        @Override
+        public int hashCode() {
+            return cls.hashCode() + name.hashCode() + Arrays.hashCode(paramTypes);
+        }
+    }
+
+    /* ********************************************************************** *
+     *                                                                        *
+     * Proxy interfaces                                                       *
+     *                                                                        *
+     * ********************************************************************** */
+
+    private interface TextLayout
+    {
+        static final int TYPE_TEXT           = 1 << 0;
+        static final int TYPE_UNDERLINE      = 1 << 1;
+
+        TextLine[] getLines();
+        int getLineIndex(float y);
+        int getCharCount();
+
+        HitInfo getHitInfo(float x, float y);
+        PathElement[] getCaretShape(int offset, boolean isLeading, float x, float y);
+        PathElement[] getRange(int start, int end, int type, float x, float y);
+    }
+
+    private interface TextLine
+    {
+        int getLength();
+        RectBounds getBounds();
+        int getStart();
+    }
+
+    private interface RectBounds
+    {
+        float getMinX();
+        float getMaxX();
+        float getHeight();
+    }
+
+    private interface HitInfo
+    {
+        int getCharIndex();
+        boolean isLeading();
     }
 }
