--- richtextfx/src/main/java/org/fxmisc/richtext/j9adapters/HitInfo.java
@@ -1,7 +1,0 @@
-package org.fxmisc.richtext.j9adapters;
-
-public interface HitInfo
-{
-    int getCharIndex();
-    boolean isLeading();
-}
