--- richtextfx/src/main/java/org/fxmisc/richtext/j9adapters/RectBounds.java
@@ -1,10 +1,0 @@
-package org.fxmisc.richtext.j9adapters;
-
-public interface RectBounds
-{
-    float getMinX();
-
-    float getMaxX();
-
-    float getHeight();
-}
