--- richtextfx/src/main/java/org/fxmisc/richtext/j9adapters/TextLine.java
@@ -1,10 +1,0 @@
-package org.fxmisc.richtext.j9adapters;
-
-public interface TextLine
-{
-    int getLength();
-
-    RectBounds getBounds();
-
-    int getStart();
-}
